+++ ContentPatcher/docs/release-notes.md
@@ -11,6 +11,7 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 -->
 ## Upcoming release
 * Added support for [config UI sections](author-guide/config.md#display-options) (thanks to Shockah!).
+* Fixed content packs reloaded unnecessarily on startup.
 
 ## 1.25.0
 Released 27 February 2022 for SMAPI 3.13.0 or later. See the [release highlights](https://www.patreon.com/posts/63137910).